updt pooling operator interfaces

+++ oneflow/core/common/shape_fns.cpp
@@ -2,26 +2,26 @@
 
 namespace oneflow {
 
-void GetWindowedOutputSizeVerboseV2(int64_t input_size, int64_t filter_size,
-                                    int64_t dilation_rate, int64_t stride,
+void GetWindowedOutputSizeVerboseV2(int32_t input_size, int32_t filter_size,
+                                    int32_t dilation_rate, int32_t stride,
                                     const std::string& padding_type,
-                                    int64_t* output_size,
-                                    int64_t* padding_before,
-                                    int64_t* padding_after) {
+                                    int32_t* output_size,
+                                    int32_t* padding_before,
+                                    int32_t* padding_after) {
   if (stride <= 0) { LOG(FATAL) << "Stride must be > 0, but got " << stride; }
   if (dilation_rate < 1) {
     LOG(FATAL) << "Dilation rate must be >= 1, but got " << dilation_rate;
   }
 
   // See also the parallel implementation in GetWindowedOutputSizeFromDimsV2.
-  int64_t effective_filter_size = (filter_size - 1) * dilation_rate + 1;
+  int32_t effective_filter_size = (filter_size - 1) * dilation_rate + 1;
   if (padding_type == "valid") {
     *output_size = (input_size - effective_filter_size + stride) / stride;
     *padding_before = *padding_after = 0;
   } else if (padding_type == "same") {
     *output_size = (input_size + stride - 1) / stride;
-    const int64_t padding_needed = std::max(
-        static_cast<int64_t>(0),
+    const int32_t padding_needed = std::max(
+        static_cast<int32_t>(0),
         (*output_size - 1) * stride + effective_filter_size - input_size);
     // For odd values of total padding, add more padding at the 'right'
     // side of the given dimension.
@@ -35,30 +35,30 @@ void GetWindowedOutputSizeVerboseV2(int64_t input_size, int64_t filter_size,
   }
 }
 
-void GetWindowedOutputSizeVerbose(int64_t input_size, int64_t filter_size,
-                                  int64_t stride,
+void GetWindowedOutputSizeVerbose(int32_t input_size, int32_t filter_size,
+                                  int32_t stride,
                                   const std::string& padding_type,
-                                  int64_t* output_size, int64_t* padding_before,
-                                  int64_t* padding_after) {
+                                  int32_t* output_size, int32_t* padding_before,
+                                  int32_t* padding_after) {
   GetWindowedOutputSizeVerboseV2(input_size, filter_size,
                                  /*dilation_rate=*/1, stride, padding_type,
                                  output_size, padding_before, padding_after);
 }
 
-void GetWindowedOutputSize(int64_t input_size, int64_t filter_size,
-                           int64_t stride, const std::string& padding_type,
-                           int64_t* output_size, int64_t* padding_size) {
-  int64_t padding_after_unused;
+void GetWindowedOutputSize(int32_t input_size, int32_t filter_size,
+                           int32_t stride, const std::string& padding_type,
+                           int32_t* output_size, int32_t* padding_size) {
+  int32_t padding_after_unused;
   GetWindowedOutputSizeVerbose(input_size, filter_size, stride, padding_type,
                                output_size, padding_size,
                                &padding_after_unused);
 }
 
-void GetWindowedOutputSizeV2(int64_t input_size, int64_t filter_size,
-                             int64_t dilation_rate, int64_t stride,
+void GetWindowedOutputSizeV2(int32_t input_size, int32_t filter_size,
+                             int32_t dilation_rate, int32_t stride,
                              const std::string& padding_type,
-                             int64_t* output_size, int64_t* padding_size) {
-  int64_t padding_after_unused;
+                             int32_t* output_size, int32_t* padding_size) {
+  int32_t padding_after_unused;
   GetWindowedOutputSizeVerboseV2(input_size, filter_size, dilation_rate, stride,
                                  padding_type, output_size, padding_size,
                                  &padding_after_unused);
